feat: bootstrap stack creation from node
