+++ cdk-bootstrapper/utils/safeAssertString.ts
@@ -1,0 +1,5 @@
+function safeAssertString(val: unknown): asserts val is string {
+  if (typeof val !== 'string') throw new Error('Expected a string.');
+}
+
+export default safeAssertString;
